add Request Validation

+++ app/Http/Controllers/EventController.php
@@ -8,6 +8,9 @@ use Illuminate\Http\Request;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Carbon\Carbon;
 use App\Http\Transformer\GetEventsTransformer;
+use App\Http\Requests\CreateEventRequest;
+use App\Http\Requests\UpdateEventRequest;
+use Illuminate\Support\Facades\DB;
 
 class EventController extends Controller{
 
@@ -21,8 +24,11 @@ class EventController extends Controller{
         $response = $transformer->transform($events);
         return response()->json($response);
     }
-    public function create(Request $request)
-    {
+    public function create(CreateEventRequest $request)
+{
+    DB::beginTransaction();
+
+    try {
         $event = new Event();
         $event->name = $request->name;
         $event->trigger_time = Carbon::parse($request->trigger_time);
@@ -37,32 +43,57 @@ class EventController extends Controller{
         }
 
         $event->eventNotifyChannels()->saveMany($eventNotifyChannels);
-
+        DB::commit();
 
         return response()->json($event);
+    } catch (\Exception $e) {
+        DB::rollBack();
+        return response()->json(['error' => 'An error occurred while creating the event.'], 500);
     }
+}
 
-    public function update($id, Request $request)
-    {
-        $updateEvent = Event::where('id', $id)->first();
-        $updateEvent->name = $request->name;
-        $updateEvent->trigger_time = Carbon::parse($request->trigger_time);
-        $updateEvent->save();
+public function update($id, UpdateEventRequest $request)
+{
+    DB::beginTransaction();
 
-        $updateEvent->eventNotifyChannels()->delete();
+    try {
+        $updateEvent = Event::where('id', $id)->firstOrFail();
 
-        $eventNotifyChannels = [];
-        foreach ($request->event_notify_channels as $eventNotifyChannelId) {
-            $eventNotifyChannel = new EventNotifyChannel();
-            $eventNotifyChannel->notify_channel_id = $eventNotifyChannelId;
-            $eventNotifyChannel->message = 'test';
-            $eventNotifyChannels[] = $eventNotifyChannel;
+        if ($request->has('name')) {
+            $updateEvent->name = $request->name;
         }
 
-        $updateEvent->eventNotifyChannels()->saveMany($eventNotifyChannels);
+        if ($request->has('trigger_time')) {
+            $updateEvent->trigger_time = Carbon::parse($request->trigger_time);
+        }
+
+        $updateEvent->save();
+
+        if ($request->has('event_notify_channels')) {
+            $updateEvent->eventNotifyChannels()->delete();
+
+            $eventNotifyChannels = [];
+            foreach ($request->event_notify_channels as $eventNotifyChannelId) {
+                $eventNotifyChannel = new EventNotifyChannel();
+                $eventNotifyChannel->notify_channel_id = $eventNotifyChannelId;
+                $eventNotifyChannel->message = 'test';
+                $eventNotifyChannels[] = $eventNotifyChannel;
+            }
+
+            $updateEvent->eventNotifyChannels()->saveMany($eventNotifyChannels);
+        }
+
+        DB::commit();
 
         return response()->json($updateEvent);
+    } catch (\Exception $e) {
+        DB::rollBack();
+
+        return response()->json(['error' => 'An error occurred while updating the event.'], 500);
     }
+}
+
+
     public function get($event_id)
     {
         $event = Event::find($event_id);
@@ -73,7 +104,8 @@ class EventController extends Controller{
             'event_notify_channels' => $event->eventNotifyChannels->pluck('notify_channel_id'),
         ];
         return response()->json($response);
-    }        public function delete($id, Request $request)
+    }        
+    public function delete($id, Request $request)
     {
         $deleteEvent = Event::where('id', $id)->first();
         $deleteEvent-> delete();
